command: new experimental "terraform checks" command
This is a placeholder way to inspect the check results recorded in the
most recent state snapshot, if any.

It isn't clear yet whether viewing checks in isolation is actually useful,
and so this command might eventually get removed again in favor of
integrating check reports into the main plan and apply flow, but until
then this is useful to see what check data Terraform is tracking.

One significant downside of this separate command is that it requires
direct access to remote state snapshots and cannot work with remote
operations. That might well end up being the key reason for integrating
into the main plan/apply flow rather than as a separate command, unless
we can find a different design that allows fetching check results from
a dedicated API when using the Terraform Cloud integration.