core: Don't dupe-register checkable objects for outputs during apply
Unfortunately we don't handle resource instances and output value
instances in the same way during the apply step. For resource instances we
generate a graph topology that has the instances already expanded so that
the initial graph is resource instances directly, but for output values
we just generate the same "expand" node we would use during planning,
which will therefore do its own registration of the dynamic checkable
objects during its own evaluation.